Change the way we generate proposals geozone links
As we want the proposals geozones links to point to the new combo box
within the advanced search.

We needed to remove the id attribute from the new custom geozone view as
we now render it at the same page many times. Otherwise the generated HTML
would be invalid.

No additional test are needed as the new feature behaves exactly the same
than the previous version but using the geozone filter instead of the search
by text.