Change proposals map geozone links and map image links
Now we want the geozone links and the map image areas to point
to the proposals advanced search filter by geozone.

No additional test are needed as the feature keep working the same
it was with the previous version but using the advanced search form
filter by geozone instead of the search by plain text.